[level 1] Title: 시저 암호, Time: 31.06 ms, Memory: 99.4 MB -BaekjoonHub

+++ 프로그래머스/1/12926. 시저 암호/시저 암호.java
@@ -5,8 +5,8 @@ class Solution {
         for(char c : ch){
             if((c+"").equals(" ")){
                 answer += " ";
-            }else if(((int)c + n > 90 && (int)c <= 90) || ((int)c + n > 122 && (int)c <= 122)) answer += (char)((int)c + n -26);
-            else answer += (char)((int)c+n) +"";
+            }else if((c + n > 90 && c <= 90) || (c + n > 122 && c <= 122)) answer += (char)(c + n -26);
+            else answer += (char)(c+n) +"";
            
         }
         return answer;
